I am adding a tab with two-sample t-test
I have added a basic tab for a two-sample t-test. It can use the uploaded data and enables an end user to select, what the continuous and grouping variable is. Also, the end user can determine what a definition of alternative hypothesis is and what mu of null hypothesis is.