Logic of paired tests fixed
When a paired test is selected such as paired t-test of Wilcoxon paired test, the second selected numerical variable is forced to be different than the first one, since it is now hidden. That makes sense -- it is not meaningful to compare two exactly same variables (differences are zero and p-value is 1.00).